+++ data_for_test.go
@@ -917,7 +917,7 @@ var suites = []FixtureSuite{
 	},
 	{
 		{
-			Name: "otel-cli exec with arg injection",
+			Name: "otel-cli exec with arg injection injects the traceparent",
 			Config: FixtureConfig{
 				CliArgs: []string{
 					"exec", "--endpoint", "{{endpoint}}",
@@ -928,6 +928,24 @@ var suites = []FixtureSuite{
 			Expect: Results{
 				Config:    otelcli.DefaultConfig().WithEndpoint("{{endpoint}}"),
 				CliOutput: "00-e39280f2980af3a8600ae98c74f2dabf-023eee2731392b4d-01\n",
+				SpanCount: 1,
+			},
+		},
+	},
+	{
+		{
+			Name: "otel-cli exec --tp-disable-inject returns the {{traceparent}} tag unmodified",
+			Config: FixtureConfig{
+				CliArgs: []string{
+					"exec", "--endpoint", "{{endpoint}}",
+					"--force-trace-id", "e39280f2980af3a8600ae98c74f2dabf", "--force-span-id", "023eee2731392b4d",
+					"--tp-disable-inject",
+					"--",
+					"echo", "{{traceparent}}"},
+			},
+			Expect: Results{
+				Config:    otelcli.DefaultConfig().WithEndpoint("{{endpoint}}"),
+				CliOutput: "{{traceparent}}\n",
 				SpanCount: 1,
 			},
 		},